+++ server/src/main/kotlin/br/edu/vaultroom/repository/ProductRepository.kt
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.Modifying
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
 import org.springframework.stereotype.Repository
+import java.math.BigDecimal
 import java.time.Instant
 
 @Repository
@@ -47,6 +48,6 @@ interface ProductRepository : JpaRepository<Product, Long> {
     fun subtotalByVaultId(
         @Param("vaultId") vaultId: Long,
         @Param("type") type: Long?
-    ): Double?
+    ): BigDecimal?
 }
